removed deleted sector editing
Deleted sectors had their floor texture named "_REMOVED" as I was having trouble cleanly removing the sectors. This is messy, and a nodebuilder will remove any unused sectors automatically, so it was also unnecessary. 
It may be an idea to eventually remove these combined sectors properly, but this works fine for now.